+++ app/Http/Controllers/Api/V1/UserController.php
@@ -9,7 +9,6 @@ use App\Http\Requests\Api\V1\UpdateUserRequest;
 use App\Http\Resources\V1\UserResource;
 use App\Models\User;
 use App\Policies\V1\UserPolicy;
-use Illuminate\Auth\Access\AuthorizationException;
 use Illuminate\Database\Eloquent\ModelNotFoundException;
 
 class UserController extends ApiController
@@ -34,13 +33,13 @@ class UserController extends ApiController
     {
         try {
 
-            $this->isAble('store', User::class);
+            if ($this->isAble('store', User::class)) {
+                return new UserResource(User::create($request->mappedAttributes()));
+            };
 
-            return new UserResource(User::create($request->mappedAttributes()));
+            return $this->error("You are not authorized to create this user", 403);
         } catch (ModelNotFoundException $e) {
             return $this->error("User cannot found", 404);
-        } catch (AuthorizationException $e) {
-            return $this->error("You are not authorized to create this user", 403);
         }
     }
 
@@ -65,15 +64,16 @@ class UserController extends ApiController
         try {
             $user = User::findOrFail($user_id);
 
-            $this->isAble('update', $user);
 
-            $user->update($request->mappedAttributes());
+            if ($this->isAble('update', $user)) {
+                $user->update($request->mappedAttributes());
 
-            return new UserResource($user);
+                return new UserResource($user);
+            }
+
+            return $this->error("You are not authorized to update this user", 403);
         } catch (ModelNotFoundException $e) {
             return $this->error("User cannot found", 404);
-        } catch (AuthorizationException $e) {
-            return $this->error("You are not authorized to update this user", 403);
         }
     }
 
@@ -85,15 +85,16 @@ class UserController extends ApiController
         try {
             $user = User::findOrFail($user_id);
 
-            $this->isAble('replace', $user);
+            if ($this->isAble('replace', $user)) {
 
-            $user->update($request->mappedAttributes());
+                $user->update($request->mappedAttributes());
 
-            return new UserResource($user);
+                return new UserResource($user);
+            }
+
+            return $this->error("You are not authorized to replace this user", 403);
         } catch (ModelNotFoundException $e) {
             return $this->error("User cannot found", 404);
-        } catch (AuthorizationException $e) {
-            return $this->error("You are not authorized to replace this user", 403);
         }
     }
 
@@ -105,15 +106,16 @@ class UserController extends ApiController
         try {
             $user = User::findOrFail($user_id);
 
-            $this->isAble('delete', $user);
+            if ($this->isAble('delete', $user)) {
 
-            $user->delete();
+                $user->delete();
 
-            return $this->ok("User deleted successfully");
+                return $this->ok("User deleted successfully");
+            }
+
+            return $this->error("You are not authorized to delete this user", 403);
         } catch (ModelNotFoundException $e) {
             return $this->error("User cannot found", 404);
-        } catch (AuthorizationException $e) {
-            return $this->error("You are not authorized to delete this user", 403);
         }
     }
 }
